move and apply `handleGraphQLHandlerError` to graphql.ts
