+++ grafast/grafserv/src/middleware/graphql.ts
@@ -300,7 +300,7 @@ export function validateGraphQLBody(
   return parsed as ValidatedGraphQLBody;
 }
 
-export const makeGraphQLHandler = (
+const _makeGraphQLHandlerInternal = (
   resolvedPreset: GraphileConfig.ResolvedPreset,
   dynamicOptions: OptionsFromConfig,
   schemaOrPromise: PromiseOrDirect<GraphQLSchema> | null,
@@ -565,3 +565,67 @@ export const makeGraphQLHandler = (
     }
   };
 };
+
+export const makeGraphQLHandler = (
+  resolvedPreset: GraphileConfig.ResolvedPreset,
+  dynamicOptions: OptionsFromConfig,
+  schemaOrPromise: PromiseOrDirect<GraphQLSchema> | null,
+) => {
+  const handler = _makeGraphQLHandlerInternal(
+    resolvedPreset,
+    dynamicOptions,
+    schemaOrPromise,
+  );
+  return (
+    request: NormalizedRequestDigest,
+    graphiqlHandler?: (
+      request: NormalizedRequestDigest,
+    ) => Promise<HandlerResult | null>,
+  ) =>
+    handler(request, graphiqlHandler).catch((e) =>
+      handleGraphQLHandlerError(request, dynamicOptions, e),
+    );
+};
+
+function handleGraphQLHandlerError(
+  request: NormalizedRequestDigest,
+  dynamicOptions: OptionsFromConfig,
+  e: Error | SafeError,
+) {
+  if (e instanceof SafeError) {
+    return {
+      type: "graphql",
+      request,
+      dynamicOptions,
+      payload: {
+        errors: [
+          new GraphQLError(e.message, null, null, null, null, e, e.extensions),
+        ],
+      },
+      statusCode: e.extensions?.statusCode ?? 500,
+      // FIXME: we should respect the `accept` header here if we can.
+      contentType: APPLICATION_JSON,
+    } as HandlerResult;
+  }
+  // TODO: if a GraphQLError is thrown... WTF?
+  const graphqlError =
+    e instanceof GraphQLError
+      ? e
+      : new GraphQLError("Unknown error occurred", null, null, null, null, e);
+  // Special error handling for GraphQL route
+  console.error(
+    "An error occurred whilst attempting to handle the GraphQL request:",
+  );
+  console.dir(e);
+  return {
+    type: "graphql",
+    request,
+    dynamicOptions,
+    payload: { errors: [graphqlError] },
+    statusCode:
+      (graphqlError.extensions?.statusCode as number | undefined) ?? 500,
+    // Fall back to application/json; this is when an unexpected error happens
+    // so it shouldn't be hit.
+    contentType: APPLICATION_JSON,
+  } as HandlerResult;
+}
